+++ braindecode/models/util.py
@@ -198,7 +198,7 @@ models_mandatory_parameters: list[
     ("ATCNet", ["n_chans", "n_outputs", "n_times"], None),
     ("BDTCN", ["n_chans", "n_outputs"], None),
     ("Deep4Net", ["n_chans", "n_outputs", "n_times"], None),
-    ("DeepSleepNet", ["n_outputs"], None),
+    ("DeepSleepNet", ["n_chans", "n_outputs", "n_times"], None),
     ("EEGConformer", ["n_chans", "n_outputs", "n_times"], None),
     ("EEGInceptionERP", ["n_chans", "n_outputs", "n_times", "sfreq"], None),
     ("EEGInceptionMI", ["n_chans", "n_outputs", "n_times", "sfreq"], None),